Move SecurityMetadata into separate file
